1099: remove requirement text on borough president waive hearings popup
Because borough president's are NOT required to submit a hearing to
comply with the City Charter, the modal warning message that notified
the user of this requirement was removed. Instead, the modal now just
lists (1) the name of the rezoning and (2) a question whether the
user is sure they want to opt out of hearing.