874 fix action filter (#1101)
* 847: re-add action filter back to projects list page

Previously the action filter on the projects list page was not working
and had to be removed. The commit where it was removed is commit
d95b8e71657fe0956d2f2a3d52c88bf6eca3e5fb.

This commit re-adds the action filter, which is now working again.
The reason for the filter break is not known.

* 874: add acceptance test for action filter

This commit adds a test in filter-checkbox-test acceptance test to
assure that query parameters are changing appropriately when a user
interacts with the action filter. This test does not actually test the
full filter functionality, since that is handled in the backend.