530: add action lookup for new action codes CS and SG, update LD tooltip
Add two new actions, Substantial Compliance (CS) and POPS Signage (SG)
to lookup-action-type. These actions will now show up in the action
type filter on the projects list page, in the Actions section on the
project profile page, and their associated tooltip will be updated with
a definition of the action.

Update tooltip for Legal Document (LD) in lookup-action-type.