Update zoning resolution URLs to use Online ZR
Zoning Resolution links to DCPs website are now broken.

This commit directs users to the new Online ZR website instead, by
bringing them to the search page if the "DCP zoningresolution" value is
a zoning resolution (ex. "74-533").

Because "DCP zoningresolution" value can be "AppendixA, AppendixB, etc."
this commit hardcodes the Appendix URLs, which is a better UX
than getting sent to a search page with no results.

This is a quick fix for broken links while we figure out a better way
to directly link users to the exact Zoning Resolution.