1019: Fix field validation bug in recommendation form
Previously, a user selection of the recommendation options `Disapproved`
and `Received after Clock Expired` would cause the Votes in Favor, Votes
Against, Abstain, and Total Members fields to automatically validate.
This would happen without the user interacting with those fields or
clicking the "Continue" button, which is not expected behavior.

The bug was being caused by a recommendation code error in the action
`setDispositionChangesetRec` in the `recommendations/add.js` controller.
The action was checking for three recommendation codes (e.g. 717170002)
and assuming these all to correspond to the label `Waiver of
Recommendation`. But different codes correlate to different labels
based  on the participant type (e.g. `Community Board`).

See `RECOMMENDATION_OPTIONSET_BY_PARTICIPANT_TYPE_LOOKUP` in
`recommendations/add.js` controller for these recommendation
code-to-label relationships.

In order to fix this, the if statement checking for relevant
recommendation codes in this action was broken out into two separate if
statements,each one checking for codes that correlated to a specific
participant type.