12122 Fix defaultValue for nameofpersoncompletingthisform
The Disposition model's dcpNameofpersoncompletingthisform
property was not being included in request payloads
to the backend. The fix is to remove the
`string-null-triggers-default` Ember Data transform,
and instead set up a pipe on the dispostion PATCH
endpoint to provide defaults/transform values.

Todo: standardize usage of pipes for setting default
values for incoming attributes.